\name{get.w.remove.j}
\alias{get.w.remove.j}

\title{
W as an vector 
}
\description{
To obtain W as an vector after removing the jth column from matrix W (by makeing the jth column as an Zero vector). 
}
\usage{
get.w.remove.j(w, j)
}

\arguments{
  \item{w}{
The W matrix.
}
  \item{j}{
The column index number of the W matrix.
}
}

\value{
A vector derived from W.
}

\keyword{ internal }

